Split the debug-id script into a thin entry and a tested module
The CLI block was the only uncovered code in the file, because a test
imports the module and coverage then measures the entry wiring too.
`scripts/check-comments.ts` and `scripts/precommit.ts` already show the
shape: an entry file holds the argument parsing, and the logic it calls
lives in its own module.

Also make `jsSiblingPath` idempotent, so a bundle that is already
JavaScript keeps its name instead of gaining a second extension.
