+++ src/shared/request-trace.ts
@@ -22,10 +22,7 @@ export type RequestTrace = {
 const requestTrace = createScopedValue<RequestTrace | null>(() => null);
 
 /** Record the request being served, for as long as it is being served. */
-export const runWithRequestTrace = <T>(
-  request: Request,
-  fn: () => Promise<T>,
-): Promise<T> => {
+export const runWithRequestTrace = <T>(request: Request, fn: () => T): T => {
   const url = new URL(request.url);
   return requestTrace.run(
     { host: url.host, method: request.method, route: redactPath(url.pathname) },
@@ -36,20 +33,26 @@ export const runWithRequestTrace = <T>(
 /** The request being served, or null when nothing is being served. */
 export const getRequestTrace = (): RequestTrace | null => requestTrace.read();
 
+/** Read one fact off the request being served, or null when none is. */
+const fromTrace =
+  <T>(read: (trace: RequestTrace) => T) =>
+  (): T | null => {
+    const trace = getRequestTrace();
+    return trace ? read(trace) : null;
+  };
+
 /**
  * The route name an error report is grouped under, or null outside a request.
  * Reads like the request log line: `GET /admin/listings/[id]`.
  */
-export const getTracedRoute = (): string | null => {
-  const trace = getRequestTrace();
-  return trace && `${trace.method} ${trace.route}`;
-};
+export const getTracedRoute: () => string | null = fromTrace(
+  (trace) => `${trace.method} ${trace.route}`,
+);
 
 /**
  * The public URL an error report happened on, with every secret removed. The
  * query string is dropped whole, because it carries tokens on some routes.
  */
-export const getTracedUrl = (): string | null => {
-  const trace = getRequestTrace();
-  return trace && `https://${trace.host}${trace.route}`;
-};
+export const getTracedUrl: () => string | null = fromTrace(
+  (trace) => `https://${trace.host}${trace.route}`,
+);
